classes: added flatpak-image.bbclass.
Added a bbclass for images with flatpak support. It defines a flatpak
image feature which pulls in the flapak package group, which in turn
has the necessary runtime flatpak supporting bits. If the flatpak
distro feature is enabled, this will also inherit flatpak-repo thus
enabling flatpak repo generation for the image.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>