recipes-*/*.bbappend: reshuffled recipe fixups/additions.
Reshuffled our directory hierarchy (mostly for .bbappends) to
reflect the corresponding hierarchy of the originating layers
of the fixed up/extended recipes.

As a side-effect, I had to add new (native-extending) bbappends
for gnupg and a few of its dependencies (libksba, nth, pinentry).
I'm quite baffled as to why these were not necessary earlier...

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>