layer.conf: declare variable for populating image with repo data.
Declare a variable (FLATPAK_APP_REPOS) for defining a set of
repositories (using a per-repo .url and .key file) that will
be known and monitored by devices running the image. By default
this is set to empty in the layer configuration file (resulting
in no monitored repos), which should be overridden in the distro
or local.conf. Additionally to the .url and .key files, a user
needs to be declared for each repository using the passwd and
group fragments ${TOPDIR}/conf/flatpak-{passwd,group}.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>