flatpak-image.bbclass: define flatpak-{runtime,sdk} image variants.
The runtime image variant pulls in all the necessary runtime bits to
add support for installing and running flatpak-based applications.

The sdk variant additionally pulls in the necessary compile-time
bits for compiling applications and exporting them to flatpak
repositories. A flatpak repository (branch) will be automatically
populated with this image during image creation. This repository
(branch) can then be used to generate flatpaks for the corresponding
runtime image variant.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>