flatpak-repo.bbclass,populate-repo.sh: split out repo export task.
Split repo exporting to archive-z2 format from the creation of
the basic repo in bare-user format. The primary reason for the
split is that, in principle, the basic repo creation could be
sstate-cached (we use a dedicated per-build/image repo) while
exporting to archive-z2 cannot (we usually use a single repo
cumulating all the builds). For now sstate-caching is disabled
altogether.

Also changed the branches we use, omitting the old build branch,
and using an image version branch, a distro version branch, and
a rolling branch. I expect this to still change, once the other
annoyances are sorted out...

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>