flatpak-repo.bbclass: default to runtime, ignore initramfs.
Default to 'runtime' flatpak type instead of 'none'. This should
allow us to build more reasonable flatpak repositories for images
that are not flatpak-enabled (mostly for dropping into a shell of
such images for testing).

Also we now never build a flatpak repository of anything that
looks like an initramfs image.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>