glib-networking: nuke ${libdir}/pkgconfig if it has no *.pc files.
relocatable.bbclass:relocatable_native_pcfiles() does not handle
gracefully ${libdir}/pkgconfig if it does not have any .pc files.
Since the patch to fix this in relocatable.bbclass is not welcome
upstream (on the basis of very few packages being broken in that
way) we need to fix things up here in one of the known offending
ones.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>