layer.conf: bbmask fuse from meta-openembedded/meta-filesystems for now.
We've been only testing ostree/flatpak with fuse@flatpak. After the
latest rebase on intel/intel-iot-refkit/master fuse@flatpak gets
shadowed by the one in meta-openembedded/meta-filesystems and the build
fails with a painfully unhelpful error message from the supported
recipe check (along the lines of fuse unsupported, would be supported
in flatpak, fuise@flatpak).

bbamsk fuse in the filesystems layer to get rid of this for now. We
can check later if the one there is sufficient for our/ostrees needs
and switch if it is.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>