flatpak-applications.inc: disable gobject-introspection data.
Disable gobject-introspection data generation. The necessary patch
to make that work with usrmerge enabled was not accepted upstream.

Since it was a patch to a bbclass we cannot simply fix it up with
a bbappend. We would either need to fork the containing layer
(openembedded-core) or add a layer with a higher priority and
override the whole class there. None of those are too compelling
considering the fact that we don't need introspection data ATM.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>