refkit-ostree: update to latest version (refkit-ostree-update in C).
Update to the latest refkit-ostree-upgrade version which now has
a simple update daemon written in C.

The same code has also bits for practically all the necessary
functionality we currently have as shell scripts for the initramfs
ostree module. The next step will be gutting out the module and
replace most of the equivalent shell scriptlets with calls to the
C-based implementation.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>