lib: oeqa: remove __init__.py in root qa dir and runtime
Documentation states that the __init__.py should only
exist the "runtime" directory. A bit missleading documentation
though as it can only exist under other directories inside
the runtime directory. Otherwise we override the packages that
are in poky, producing errors like this:

   Exception: ImportError: No module named 'oeqa.core'

Fixed by only keeping the __init__.py file in the same directory
as the actually test file.

Signed-off-by: Mirza Krak <mirza.krak@endian.se>