Update Flatpak to handle /var/tmp being a symlink
Flatpaks may request /var/tmp be bind-mounted from the host.
If mounts are symlinks then flatpak attempts to create the symlinks
and bind-mount the target directory instead.

This doesn't work for the /var/tmp → /var/volatile/tmp symlink in Yocto
because flatpak already created that directory and mounted something there.

This is fixed by adding /var/tmp to a blacklist that already exists for /tmp
because it has the same problem, and has the effect of bind-mounting
the host's /var/volatile/tmp directory into the container's /var/tmp.